backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
supabase>=2.15.0
python-dotenv>=1.1.0
python-multipart>=0.0.20
pyjwt>=2.10.0
httpx>=0.28.0
